All the Exploratory Model Evaluation is complete
There is no new Training or Recognition Model to explore.

Currectly Promishing models are LinearSVM and Euclidian Distance. The custom decision parameter for evaluating a face as correct for linearSVM is positive value greater than a arbitary number 2 and for EuclidianDistance the shortest distance less than a arbitary number 50->0.5

KNN clasifier performance is worsed. it is  showing very high false positive with confidace of 100%. This model should be depricated. This model should improve if equal number of training faces are present for each person.

automated Testing program will be implimented next and with that the timing an accuracy will be predicted. confiusion matrix is prefarable.